feat: modify data type

See merge request s13-fintech-finance-sub1/S13P21A407!74

+++ pocketc_backend/src/main/java/com/ssafy/pocketc_backend/domain/transaction/dto/request/TransactionCreateReqDto.java
@@ -5,10 +5,10 @@ import java.time.LocalDateTime;
 public record TransactionCreateReqDto(
         Integer userId,
         LocalDateTime date,
-        Integer amount,
+        Long amount,
         String merchantName
 ) {
-    public static TransactionCreateReqDto of(Integer userId, LocalDateTime date, Integer amount, String merchantName) {
+    public static TransactionCreateReqDto of(Integer userId, LocalDateTime date, Long amount, String merchantName) {
         return new TransactionCreateReqDto(userId, date, amount, merchantName);
     }
 }
